refactor(storage): extract UserPlaybookStoreMixin from PlaybookMixin
Behavior-preserving verbatim move of the 18 user-playbook public methods
(+ 2 SQLite privates) out of the monolithic PlaybookMixin into a new
UserPlaybookStoreMixin, for the SQLite backend and the storage_base ABC.

Each backend gains a playbook/ package (playbook/_user.py) re-exported via
playbook/__init__.py; the composition base lists add UserPlaybookStoreMixin
alongside the now-shrunken PlaybookMixin so the public surface is unchanged.

Shared module-level helpers (_emit_hard_delete_playbook, _build_tags_sql)
stay in _playbook.py and are imported by _user.py. The user-only
_emit_supersede_user_playbook helper moves into _user.py; the atomicity
characterization test's monkeypatch target is repointed accordingly.

First Phase-B extraction; sets the conventions reused by later buckets.
